promoting size from int to long, http://datahub.io/dataset/hbz_unioncatalog/resource/15dc7d28-91ba-4843-bbb7-9b1ffcb9d4d3 had a size bigger than int and it choked GSON.

+++ src/org/ckan/Resource.java
@@ -20,7 +20,7 @@ public class Resource{
     private String cache_last_updated;
     private String package_id;
     private String webstore_last_updated;
-    private int size;
+    private long size;
     private int position;
     private String resource_type;
     private String last_modified;
@@ -118,11 +118,11 @@ recent: 3
         return webstore_last_updated;
     }
 
-    public void setSize(int size) {
+    public void setSize(long size) {
         this.size = size;
     }
 
-    public int getSize() {
+    public long getSize() {
         return size;
     }
